# docs/api/openapi.yaml
openapi: 3.0.0
paths:
...
  /jobs:
    get:
      operationId: JobsController_list
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: pageSize
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 25
            type: number
      responses:
        '200':
          description: ''
      tags:
        - jobs
      security:
        - bearer: []
    post:
      operationId: JobsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobDto'
      responses:
        '201':
          description: ''
      tags:
        - jobs
      security:
        - bearer: []
  /jobs/{id}:
    get:
      operationId: JobsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - jobs
      security:
        - bearer: []
    patch:
      operationId: JobsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateJobDto'
      responses:
        '200':
          description: ''
      tags:
        - jobs
      security:
        - bearer: []
    delete:
      operationId: JobsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - jobs
      security:
        - bearer: []
  /jobs/{id}/start:
    post:
      operationId: JobsController_start
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      tags:
        - jobs
      security:
        - bearer: []
  /jobs/{id}/complete:
    post:
      operationId: JobsController_complete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      tags:
        - jobs
      security:
        - bearer: []
  /jobs/{id}/assignments:
    post:
      operationId: JobsController_createAssignment
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobAssignmentDto'
      responses:
        '201':
          description: ''
      tags:
        - jobs
      security:
        - bearer: []
  /jobs/{id}/services:
    post:
      operationId: JobsController_createJobService
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobServiceDto'
      responses:
        '201':
          description: ''
      tags:
        - jobs
      security:
        - bearer: []
...
  /invoices:
    get:
      operationId: InvoicesController_list
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: pageSize
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 25
            type: number
      responses:
        '200':
          description: ''
      tags:
        - invoices
      security:
        - bearer: []
    post:
      operationId: InvoicesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceDto'
      responses:
        '201':
          description: ''
      tags:
        - invoices
      security:
        - bearer: []
  /invoices/{id}:
    get:
      operationId: InvoicesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - invoices
      security:
        - bearer: []
    patch:
      operationId: InvoicesController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInvoiceDto'
      responses:
        '200':
          description: ''
      tags:
        - invoices
      security:
        - bearer: []
  /invoices/{id}/line-items:
    post:
      operationId: InvoicesController_createLineItem
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceLineItemDto'
      responses:
        '201':
          description: ''
      tags:
        - invoices
      security:
        - bearer: []
  /payments:
    get:
      operationId: PaymentsController_list
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: pageSize
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 25
            type: number
      responses:
        '200':
          description: ''
      tags:
        - payments
      security:
        - bearer: []
    post:
      operationId: PaymentsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentDto'
      responses:
        '201':
          description: ''
      tags:
        - payments
      security:
        - bearer: []
  /payments/{id}:
    get:
      operationId: PaymentsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - payments
      security:
        - bearer: []
...
info:
  title: DOS API
  description: >-
    Digital Operations System backend API. Every route except the ones marked
    @Public() requires a Bearer access token scoped to the caller's active
    organization.
  version: 0.1.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateJobDto:
      type: object
      properties:
        clientId:
          type: string
        serviceAddressId:
          type: string
        scheduledStart:
          type: string
        scheduledEnd:
          type: string
        recurrenceRule:
          type: string
          description: RRULE string (RFC 5545) for recurring jobs.
        notes:
          type: string
      required:
        - clientId
    UpdateJobDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - DRAFT
            - SCHEDULED
            - IN_PROGRESS
            - COMPLETED
            - CANCELLED
        serviceAddressId:
          type: string
        scheduledStart:
          type: string
        scheduledEnd:
          type: string
        actualStart:
          type: string
          description: When work actually started — distinct from the scheduled time.
        actualEnd:
          type: string
          description: When work actually ended — distinct from the scheduled time.
        notes:
          type: string
    CreateJobAssignmentDto:
      type: object
      properties:
        membershipId:
          type: string
          description: The OrganizationMembership id of the staff member to assign.
      required:
        - membershipId
    CreateJobServiceDto:
      type: object
      properties:
        serviceId:
          type: string
        quantity:
          type: number
        notes:
          type: string
      required:
        - serviceId
        - quantity
...
    CreateInvoiceDto:
      type: object
      properties:
        clientId:
          type: string
        issueDate:
          type: string
        dueDate:
          type: string
      required:
        - clientId
        - issueDate
        - dueDate
    UpdateInvoiceDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - DRAFT
            - SENT
            - PAID
            - OVERDUE
            - VOID
        dueDate:
          type: string
    CreateInvoiceLineItemDto:
      type: object
      properties:
        jobId:
          type: string
        serviceId:
          type: string
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
      required:
        - description
        - quantity
        - unitPrice
    CreatePaymentDto:
      type: object
      properties:
        invoiceId:
          type: string
        amount:
          type: number
        method:
          type: string
          enum:
            - CASH
            - CHECK
            - CARD
            - BANK_TRANSFER
            - OTHER
        referenceNumber:
          type: string
        notes:
          type: string
      required:
        - invoiceId
        - amount
        - method